requirements.txt
anyio==4.8.0
certifi==2025.1.31
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pipreq==0.4
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
setuptools==75.8.0
sgmllib3k==1.0.0
sniffio==1.3.1
telegram==0.0.1
typing_extensions==4.12.2
